requirements.txt
allure-pytest==2.8.31
allure-python-commons==2.8.31
jsonpath==0.82
py==1.9.0
pytest==6.0.1
requests==2.24.0
xlrd==1.2.0
XlsxWriter==1.3.3
xlutils==2.0.0
xlwt==1.3.0
pytest-xdist==2.5.0
pytest-html==3.1.1